Starting Edge Rework
The way I had it set up before doesn't work for actually detecting edges
that are the same, I need to go about it differently.

My thought is to have an array of 40 pairs of indexes/sides that
connect. For the original puzzle this is easy. For solutions, I would
need to go through the same way I made the original array to make
another one, this time accounting for rotations with which side is being
added, and then compare the two arrays to count the number of shared
elements