Fixing Pair Checking
There were multiple bugs in the pair checking function, most appear to
be fixed now. The main issue I seem to be having is now getting the
sides connection to read out correctly. The original solution will
correctly give (40) as the number of original indexes that are next to
each other, but will only give (20) to the number of side connections,
which should also be (40).