Playing Around with Genetic Algorithm
Just tweaking some things. This is a commit in order to start working on
the pair check, it appears that it is not reporting the edge connections
correctly and that's messing with things.

Edges that aren't on the same side as the original don't seem to be
counted at all, so the genetic alogirthm ends up finding a bunch of
solutions that just swap the left/right sides, and top/bottom sides.