Too Many Updates to List
I did a poor job maintaining version control with this one. However, I'm
starting a new methodology: only checking puzzles that have a
configuration that allows for at least (1) alternative edge solution
that contains no indexes touching that touch in the original
configuration. If I'm going to find a configuration that has no
sides/indexes the same between the two valid configurations, this is a
must. From there, I can only mutate the center connections to see if
there are any possible solutions with that unique edge solution