Pulled Edge Solving Out of findValidSolutions
In the interest of making the code a bit more understandable in case I
dump it for a few weeks and come back like I did this last time, I
pulled out the section of code that was figuring out all valid edge
solutions into its own function. This should also help with continuing
to pursue the idea of only testing puzzles that have a fully unique edge
solution. When searching for those solutions, I can store the valid edge
solutions since those are what I'm checking, and skip that step for
every center iteration beyond it since the edge solution won't be
changing.