TEST: Array vs Nested Loops for Forming Triples
I wanted to test the speed of triple nested loops (current solution) vs.
an array of all valid combinations that don't share any indexes. The
triple nested loops was ~10X faster, and certainly because the nested
loops allows a whole number to be skipped if it doesn't connect with the
previous piece, whereas the array has no way to skip large chunks like
that.