Starting Testing for New Edge Generation
The current implementation starts with an array where each unique
connection number has 2 elements, and adds random elements until 40 is
reached. This leads to situations where the edge won't even have other
solutions (there aren't at least 2 of a connection on the edge to swap).
So, I'm looking into a way to generate the edge first, and go from
there.

The idea is the same as the total puzzle generation is now: calculate
how many connections must/can be in the edge, pick a random number
between the two (inclusive), and then randomly increment each of those
elements until the total (16) is reached. This gives the edge (almost)
guaranteed ability to have multiple solutions, which is needed for
getting a totally novel second solution.

The edge could still not have another solution if the number of PIECES
that have the connection is < 2, or if multiple corners have the same
connection and there aren't enough middle-edge pieces to swap between
them.