Fixed minUniqueConnectionsInEdge Calculation
The old calculation wasn't taking into account how many unique edges the
center would be able to fill in, so it was giving numbers that would
possibly result in there not being enough center connections left to get
up to numUniqueConnections. Readjusted the formula, and now the numbers
are making sense.