Added piece_contains Function
Started as just a little test since the calculateValidCenterRows
function is by far the largest time consumer of the problem. I added a
bitfield to Piece that has bits set in all of the connections the piece
contains, and the function just checks if a specific connection is in
that bitfield, all with bitwise operations. Adding this to the second
and third loop of the big function, checking at rotation 0 to see if the
piece even contains the wanted side. This way, it can skip 3 extra
calculations (the 3 other rotations) if the piece doesn't contain the
connection dictated by the last piece.

Running the profiler:
Initial program: 6.33 seconds in center function, 14.79 total seconds
piece_contains: 4.55 seconds in center function, 11.39 total seconds
The extra time was gained from also stopping some of the other functions
in the loop from having to be called as often.