Updated CenterRow Function to Verify Edges
Since I am gien the edgeSolution as a part of that function, I might as
well check to see if the (potentially) valid row is actually fitting in
a row of the edgeSolution. Before, if it had a valid left and valid
right, it went through. Now, I have it checking to see if the valid
left/right line up in the same row or else it gets tossed. Minor speed
improvement.