Returned to Validating Center Rows
Did very very little checking, and it seems like both the validating and
not are creating the same number of center solutions, so I'm truly not
sure why it's changing the outcome of the test. I see that when using
valid, the function itself is called waaayyy less, so I'm investigating
that now!